mb/hp: Add Pro 3500 series (Sandy/Ivy Bridge)
This is another readily available (used market) system.
Based on autoport.

All peripherals should work.
Automatic fan control as well as S3 are working.
The board was tested to boot Linux and Windows. EHCI debug is untested.
When using MrChromebox edk2 with secure boot build in, the board will
hang on each boot for about 20 seconds before continuing.

There are some quirks for doing the first flash, see the documentation.

Change-Id: Idf793fe915096cf2553572964faec5c7f8526b9a
Signed-off-by: Joel Linn <jl@conductive.de>